pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "J-Stuff", email = "jstuff@j-stuff.net"},
]
dependencies = [
    "Red-DiscordBot>=3.5.5",
    "requests>=2.31.0",
    "aiodns==3.0.0",
    "aiohttp==3.8.5",
    "aiohttp-json-rpc==0.13.3",
    "aiosignal==1.3.1",
    "apsw==3.43.1.0",
    "async-timeout==4.0.3",
    "attrs==23.1.0",
    "babel==2.12.1",
    "brotli==1.1.0",
    "cffi==1.15.1",
    "charset-normalizer==3.2.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "contextlib2==21.6.0",
    "discord-py==2.3.2",
    "frozenlist==1.4.0",
    "idna==3.4",
    "importlib-metadata==6.8.0",
    "markdown==3.4.4",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "multidict==6.0.4",
    "orjson==3.9.7",
    "packaging==23.1",
    "platformdirs==3.10.0",
    "psutil==5.9.5",
    "pycares==4.3.0",
    "pycparser==2.21",
    "pygments==2.16.1",
    "python-dateutil==2.8.2",
    "pytz==2023.3.post1",
    "pyyaml==6.0.1",
    "rapidfuzz==3.3.0",
    "red-commons==1.0.0",
    "Red-DiscordBot==3.5.5",
    "red-lavalink==0.11.0",
    "rich==13.5.2",
    "schema==0.7.5",
    "six==1.16.0",
    "typing-extensions==4.7.1",
    "yarl==1.9.2",
    "zipp==3.16.2",
]
requires-python = "==3.8.10"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
